move: 디바이스 서비스 관련 이벤트 처리 로직 디바이스 이벤트 핸들러로 이동
