+++ src/main/java/com/nexters/goalpanzi/application/device/event/handler/DeviceSubscriptionEventHandler.java
@@ -4,6 +4,8 @@ import com.nexters.goalpanzi.application.auth.event.LoginEvent;
 import com.nexters.goalpanzi.application.device.DeviceSubscriptionService;
 import com.nexters.goalpanzi.application.device.event.UpdateDeviceTokenEvent;
 import com.nexters.goalpanzi.application.device.event.UpdatePushActivationStatusEvent;
+import com.nexters.goalpanzi.application.mission.event.SubscribeToMissionEvent;
+import com.nexters.goalpanzi.application.mission.event.UnsubscribeFromMissionEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -24,7 +26,7 @@ public class DeviceSubscriptionEventHandler {
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     void handleUpdateDeviceTokenEvent(final UpdateDeviceTokenEvent event) {
-        if (event.deprecatedDeviceToken() != null) {
+        if (event.isTokenDeprecated()) {
             deviceSubscriptionService.unsubscribeFromMyMissions(event.memberId(), event.deviceId(), event.deprecatedDeviceToken());
         }
         deviceSubscriptionService.subscribeToMyMissions(event.memberId(), event.deviceId());
@@ -52,4 +54,17 @@ public class DeviceSubscriptionEventHandler {
         log.info("Handled LoginEvent for memberId: {}", event.memberId());
     }
 
+    @Async
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    void handleSubscribeMissionEvent(final SubscribeToMissionEvent event) {
+        deviceSubscriptionService.subscribeToMission(event.memberId(), event.mission());
+        log.info("Handled SubscribeMissionEvent for memberId: {} and missionId: {}", event.memberId(), event.mission().getId());
+    }
+
+    @Async
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    void handleUnsubscribeFromMissionEvent(final UnsubscribeFromMissionEvent event) {
+        deviceSubscriptionService.unsubscribeFromMission(event.missionId());
+        log.info("Handled UnsubscribeMissionEvent from missionId: {}", event.missionId());
+    }
 }
