+++ src/main/java/com/nexters/goalpanzi/application/firebase/event/handler/PushMessageEventHandler.java
@@ -1,11 +1,8 @@
 package com.nexters.goalpanzi.application.firebase.event.handler;
 
-import com.nexters.goalpanzi.application.device.DeviceSubscriptionService;
 import com.nexters.goalpanzi.application.firebase.TopicGenerator;
 import com.nexters.goalpanzi.application.mission.event.CompleteMissionEvent;
 import com.nexters.goalpanzi.application.mission.event.JoinMissionEvent;
-import com.nexters.goalpanzi.application.mission.event.SubscribeToMissionEvent;
-import com.nexters.goalpanzi.application.mission.event.UnsubscribeFromMissionEvent;
 import com.nexters.goalpanzi.infrastructure.firebase.PushMessageSender;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,8 +21,6 @@ import static com.nexters.goalpanzi.domain.firebase.PushMessage.MISSION_JOINED;
 @RequiredArgsConstructor
 @Component
 public class PushMessageEventHandler {
-
-    private final DeviceSubscriptionService deviceSubscriptionService;
 
     private final PushMessageSender pushMessageSender;
 
@@ -58,19 +53,5 @@ public class PushMessageEventHandler {
                 topic
         );
         log.info("Handled CompleteMissionEvent for missionId: {}", event.missionId());
-    }
-
-    @Async
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    void handleSubscribeMissionEvent(final SubscribeToMissionEvent event) {
-        deviceSubscriptionService.subscribeToMission(event.memberId(), event.mission());
-        log.info("Handled SubscribeMissionEvent for memberId: {} and missionId: {}", event.memberId(), event.mission().getId());
-    }
-
-    @Async
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    void handleUnsubscribeFromMissionEvent(final UnsubscribeFromMissionEvent event) {
-        deviceSubscriptionService.unsubscribeFromMission(event.missionId());
-        log.info("Handled UnsubscribeMissionEvent from missionId: {}", event.missionId());
     }
 }
